feat: typeahead add async example
