+++ example/storybook/stories/components/composites/Typeahead/UsingComponentWithRenderItem.tsx
@@ -1,51 +1,47 @@
 import React from 'react';
-import { Typeahead, Box, Text, Icon, Heading, useColorMode } from 'native-base';
+import { Typeahead, Icon, Box, Text } from 'native-base';
 
-let countries = [
-  { name: 'Afghanistan', abbr: 'AFG' },
-  { name: 'Australia', abbr: 'AUS' },
-  { name: 'India', abbr: 'IND' },
-  { name: 'USA', abbr: 'USA' },
-  { name: 'Germany', abbr: 'GER' },
-  { name: 'France', abbr: 'FRA' },
-  { name: 'Iceland', abbr: 'ICL' },
-  { name: 'Russia', abbr: 'RUS' },
-  { name: 'Japan', abbr: 'JAP' },
-  { name: 'China', abbr: 'CHI' },
-  { name: 'Denmark', abbr: 'DEN' },
-  { name: 'Norway', abbr: 'NOR' },
+const animals = [
+  { key: 1, value: 'Aardvark' },
+  { key: 2, value: 'Kangaroo' },
+  { key: 3, value: 'Snake' },
+  { key: 4, value: 'Pikachu' },
+  { key: 5, value: 'Tiger' },
+  { key: 6, value: 'Godzilla' },
 ];
 
 export function UsingComponentWithRenderItem() {
-  const { colorMode } = useColorMode();
   return (
-    <Box>
-      <Heading>Typeahead</Heading>
-      <Typeahead
-        mt={4}
-        size="2xl"
-        numberOfItems={6}
-        options={countries}
-        getOptionLabel={(country: any) => country.name}
-        renderItem={(item: any) => {
-          return (
+    <Typeahead
+      items={animals}
+      disabledKeys={[1]}
+      width={400}
+      onSelectionChange={console.log}
+      label="Select your favorite animal"
+      renderItem={(item: any) => {
+        return (
+          <Box flexDirection="row" justifyContent="space-between" p={4}>
             <Box
-              flex={1}
-              bg={colorMode === 'light' ? 'red.100' : 'gray.600'}
-              p={4}
+              backgroundColor="black"
+              height={10}
+              width={10}
+              borderRadius={9999}
+              justifyContent="center"
+              alignItems="center"
             >
-              <Text>{item}</Text>
+              <Text color="white">{item.value[0]}</Text>
             </Box>
-          );
-        }}
-        toggleIcon={({ isOpen }: any) => {
-          return isOpen ? (
-            <Icon name="arrow-drop-up" type="MaterialIcons" size={12} />
-          ) : (
-            <Icon name="arrow-drop-down" type="MaterialIcons" size={12} />
-          );
-        }}
-      />
-    </Box>
+            <Box>{item.value}</Box>
+          </Box>
+        );
+      }}
+      toggleIcon={({ isOpen }: any) => {
+        return isOpen ? (
+          <Icon name="arrow-drop-up" type="MaterialIcons" size={12} />
+        ) : (
+          <Icon name="arrow-drop-down" type="MaterialIcons" size={12} />
+        );
+      }}
+    />
   );
 }
