--- src/components/composites/ComboBox/index.ts
@@ -1,3 +1,0 @@
-import { ComboBox } from './ComboBox';
-
-export { ComboBox };
