--- src/components/composites/Typeahead/useTypeahead/types.tsx
@@ -1,17 +1,0 @@
-export const InputChange = '__input_change__';
-export const InputBlur = '__input_blur__';
-
-export const ItemClick = '__item_click__';
-
-export const ToggleButtonClick = '__togglebutton_click__';
-
-export const FunctionToggleMenu = '__function_toggle_menu__';
-export const FunctionOpenMenu = '__function_open_menu__';
-export const FunctionCloseMenu = '__function_close_menu__';
-
-export const FunctionSelectItem = '__function_select_item__';
-export const FunctionSetInputValue = '__function_set_input_value__';
-
-export const FunctionReset = '__function_reset__';
-export const ControlledPropUpdatedSelectedItem =
-  '__controlled_prop_updated_selected_item__';
