+++ src/components/composites/index.ts
@@ -147,6 +147,4 @@ export { AppBar } from './AppBar';
 
 export { Fab } from './Fab';
 export type { IFabProps } from './Fab';
-export { default as Typeahead, useTypeahead } from './Typeahead';
-
-export { ComboBox } from './ComboBox';
+export { Typeahead } from './Typeahead';
